Missing group keyword in call to test_group
Hello everyone,

In the check_equipartiton function the group keyword for calls to test_group with randomised or user-defined groups was not specified, therefore the group in "test_group" defaulted to None effectively giving the same, system wide, equipartition analysis for all groups. 
I added the keyword and added a check in test_group to catch DoFs with zero DoFs, because passing these DoFs resulted in errors in the trajectory analysis of the pymbar module.