Remove python 2 support, add python 3.9 support & testing.
There's still some work arounds we want to get rid of, but for now,
we're trying to avoid large code changes before we have more
testing. On the other hand, we want to avoid having to introduce more
workarounds when introducing tests. This only changes the supported
version, adds testing for 3.9, and removes __future__ imports related
to python 2. This refs #47, but more work is needed before we close
the issue.