Remove python 2 support (#80)
Python 2 is out of support since a year, and it's time to get rid of it.
This only changes the supported version, and
removes __future__ imports related to python 2. This refs #47,
but more work is needed before we close the issue.

There's still some work arounds we want to get rid of, but for now,
we're trying to avoid large code changes before we have more
testing. On the other hand, we want to avoid having to introduce more
workarounds when introducing tests.