Add some tests for physical_validation.util.kinetic_energy (#83)
This add unit tests for parts of the
physical_validation.util.kinetic_energy module, namely

* is_close(a, b, rel_tol, abs_tol)
* temperature(kin, ndof, kb)

This also hardens the two functions somewhat: Catching some inputs
which don't make sense, and adding a sensible default value for the
absolute tolerance of is_close. This also renames isclose to is_close
to fit the style.