Simplify trajectory preparation of interval estimate
The interval estimate functionality was preparing the trajectory
by calling the different parts of the preparation (equilibration,
decorrelation, tail pruning) separately, instead of calling the
combined higher-level function.

This change replaces the calls to the lower level function by a
call to the higher-level one. The regression test confirms that
the result is unchanged.

Refs #99, as this simplification (and harmonization with other
uses of the trajectory prepataion) will make the task in #99
easier.