Allow KE equilibration check to run without observable object
KE equilibration doesn't need an observable object as input, but
uses it to store some values (which speeds up the run in case that
more than one test is run on the same simulation data). The addition
in this change allows the KE equilibration check to run even if the
input simulation data object did not initialize such an object.