Allow multidimensional trajectories to skip preparation
This switches the order of the checks during trajectory preparation:
Until now, it was first checked whether there were more than 2 dimensions
(for which preparation is not implemented), and after whether preparation
should be skipped. This is now inverted, as it isn't necessary to abort
trajectory preparation if it shouldn't be performed in the first place.