Fix kinetic energy regression plot (#196)
Updating matplotlib from 2.4.x to 2.5 leads to small but noticeable
differences in the kinetic energy plot of the corresponding
regression test.

It would be great to pin our requirements in the test environment and
use dependabot to update them, but it seems that currently conda is not
supported. I think that under these circumstances, having the tests
break when the non-pinned version changes some of the output is
actually safer than pinning them. The tests may fail for a few days
before we fix it, but at least we actually notice that a new version
changed some of the output and we can assess if that is a problem
or just a minor difference for which we will simply update the reference
files.